+++ src/Vectors.java
@@ -1,11 +1,11 @@
 
 public class Vectors {
     String name;
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 
-    public Vectors(String Name, float x_coordinate, float y_coordinate, float z_coordinate){
+    public Vectors(String Name, double x_coordinate, double y_coordinate, double z_coordinate){
         this.name = Name;
         this.x = x_coordinate;
         this.y = y_coordinate;
